[IMP] web: add sortableDragAndDrop
Add the DragAndDropSortable from nest_sortable to hoot dom helper
