+++ addons/web/static/tests/web_test_helpers.js
@@ -39,7 +39,12 @@ export {
     getDropdownMenu,
     mountWithCleanup,
 } from "./_framework/component_test_helpers";
-export { contains, defineStyle, editAce } from "./_framework/dom_test_helpers";
+export {
+    contains,
+    defineStyle,
+    editAce,
+    sortableDrag,
+} from "./_framework/dom_test_helpers";
 export {
     clearRegistry,
     getMockEnv,
